IHS-156 removed a part of fixtures to upper level
