+++ tests/unit/sdk/conftest.py
@@ -908,6 +908,114 @@ async def ipnetwork_schema() -> NodeSchemaAPI:
 
 
 @pytest.fixture
+async def ipam_ipprefix_schema() -> NodeSchemaAPI:
+    data = {
+        "name": "IPNetwork",
+        "namespace": "Ipam",
+        "default_filter": "prefix__value",
+        "display_labels": ["prefix_value"],
+        "order_by": ["prefix_value"],
+        "inherit_from": ["BuiltinIPAddress"],
+    }
+    return NodeSchema(**data).convert_api()
+
+
+@pytest.fixture
+async def simple_device_schema() -> NodeSchemaAPI:
+    data = {
+        "name": "Device",
+        "namespace": "Infra",
+        "label": "Device",
+        "default_filter": "name__value",
+        "order_by": ["name__value"],
+        "display_labels": ["name__value"],
+        "attributes": [{"name": "name", "kind": "Text", "unique": True}],
+        "relationships": [
+            {
+                "name": "primary_address",
+                "peer": "IpamIPAddress",
+                "label": "Primary IP Address",
+                "optional": True,
+                "cardinality": "one",
+                "kind": "Attribute",
+            },
+            {
+                "name": "ip_address_pool",
+                "peer": "CoreIPAddressPool",
+                "label": "Address allocator",
+                "optional": True,
+                "cardinality": "one",
+                "kind": "Attribute",
+            },
+        ],
+    }
+    return NodeSchema(**data).convert_api()
+
+
+@pytest.fixture
+async def ipam_ipprefix_data() -> dict[str, Any]:
+    return {
+        "node": {
+            "__typename": "IpamIPPrefix",
+            "id": "llllllll-llll-llll-llll-llllllllllll",
+            "display_label": "192.0.2.0/24",
+            "prefix": {
+                "is_protected": True,
+                "owner": None,
+                "source": {
+                    "__typename": "Account",
+                    "display_label": "CRM",
+                    "id": "cccccccc-cccc-cccc-cccc-cccccccccccc",
+                },
+                "value": "192.0.2.0/24",
+            },
+            "description": {
+                "is_protected": False,
+                "owner": None,
+                "source": None,
+                "value": None,
+            },
+            "member_type": {
+                "is_protected": True,
+                "owner": None,
+                "source": {
+                    "__typename": "Account",
+                    "display_label": "CRM",
+                    "id": "cccccccc-cccc-cccc-cccc-cccccccccccc",
+                },
+                "value": "address",
+            },
+            "is_pool": {
+                "is_protected": True,
+                "owner": None,
+                "source": {
+                    "__typename": "Account",
+                    "display_label": "CRM",
+                    "id": "cccccccc-cccc-cccc-cccc-cccccccccccc",
+                },
+                "value": False,
+            },
+            "ip_namespace": {
+                "properties": {
+                    "is_protected": True,
+                    "owner": None,
+                    "source": {
+                        "__typename": "Account",
+                        "display_label": "CRM",
+                        "id": "cccccccc-cccc-cccc-cccc-cccccccccccc",
+                    },
+                },
+                "node": {
+                    "id": "rrrrrrrr-rrrr-rrrr-rrrr-rrrrrrrrrrrr",
+                    "display_label": "default",
+                    "__typename": "IpamNamespace",
+                },
+            },
+        }
+    }
+
+
+@pytest.fixture
 async def address_schema() -> NodeSchemaAPI:
     data = {
         "name": "Address",
@@ -2288,6 +2396,16 @@ def query_introspection() -> str:
 
 
 @pytest.fixture
+async def mock_schema_query_ipam(httpx_mock: HTTPXMock) -> HTTPXMock:
+    response_text = (get_fixtures_dir() / "schema_ipam.json").read_text(encoding="UTF-8")
+
+    httpx_mock.add_response(
+        method="GET", url="http://mock/api/schema?branch=main", json=ujson.loads(response_text), is_reusable=True
+    )
+    return httpx_mock
+
+
+@pytest.fixture
 async def mock_query_location_batch_count(
     httpx_mock: HTTPXMock, client: InfrahubClient, mock_schema_query_01: HTTPXMock
 ) -> HTTPXMock:
@@ -2416,5 +2534,25 @@ async def nested_device_with_interfaces_schema() -> NodeSchemaAPI:
                 "kind": "Component",
             },
         ],
+    }
+    return NodeSchema(**data).convert_api()
+
+
+@pytest.fixture
+async def vlan_schema() -> NodeSchemaAPI:
+    data = {
+        "name": "VLAN",
+        "namespace": "Infra",
+        "label": "VLAN",
+        "default_filter": "name__value",
+        "order_by": ["name__value"],
+        "display_labels": ["name__value"],
+        "attributes": [
+            {"name": "name", "kind": "Text", "unique": True},
+            {"name": "vlan_id", "kind": "Number"},
+            {"name": "role", "kind": "Text", "optional": True},
+            {"name": "status", "kind": "Text", "optional": True},
+        ],
+        "relationships": [],
     }
     return NodeSchema(**data).convert_api()
